validations on select tags

+++ app/Http/Controllers/studentController.php
@@ -74,8 +74,10 @@ class studentController extends Controller
      */
     public function edit($id)
     {
+        if (Auth::user()->role != 1) {
+            return Response::HTTP_FORBIDDEN;
+        }
         $student = Student::find($id);
-
         return view('students/edit-student', compact('student'));
     }
 
@@ -88,7 +90,9 @@ class studentController extends Controller
      */
     public function update(Request $request, $id)
     {
-
+        if (Auth::user()->role != 1) {
+            return Response::HTTP_FORBIDDEN;
+        }
         $student = Student::find($id);
         $student->name = $request->get('name');
         $student->fname = $request->get('fname');
@@ -105,7 +109,9 @@ class studentController extends Controller
      */
     public function destroy($id)
     {
-
+        if (Auth::user()->role != 1) {
+            return Response::HTTP_FORBIDDEN;
+        }
 
         $student = Student::find($id);
         $student->delete();;
